+++ Services/StoreService.cs
@@ -7,8 +7,10 @@ using AutoMapper;
 using Contracts;
 using Entities.Exceptions;
 using Entities.Models;
+using Microsoft.AspNetCore.Identity;
 using Services.Contracts;
-using Shared.Data_Transfer;
+using SharedAPI.Data_Transfer;
+using SharedAPI.Request_Features;
 
 namespace Services
 {
@@ -16,26 +18,44 @@ namespace Services
     {
         private readonly IMapper mapper;
         private readonly IRepositoryManager manager;
+        private readonly UserManager<User> userManager;
 
-        public StoreService(IMapper mapper, IRepositoryManager manager)
+        public StoreService(IMapper mapper, IRepositoryManager manager, UserManager<User> userManager)
         {
             this.mapper = mapper;
             this.manager = manager;
+            this.userManager = userManager;
         }
 
-        public async Task<StoresDto> CreateStore(CreateStoreDto storeDto)
+        public async Task<StoresDto> CreateStore(string username, CreateStoreDto storeDto)
         {
-            var store = mapper.Map<Store>(storeDto);
+            //Get stores that have been created by user; if null - Create new store; else - throw exception
+            var user = await userManager.FindByNameAsync(username);
 
-            manager.stores.CreateStore(store);
-            await manager.Save();
+            var store = await manager.stores.GetStoreByOwnerId(user.Id, trackChanges: false);
+            if (store is null)
+            {
+                var newStore = mapper.Map<Store>(storeDto);
 
-            var response = mapper.Map<StoresDto>(store);
-            return response;
+                newStore.PhoneNumber = user.PhoneNumber;
+                newStore.Email = user.Email;
+                newStore.UserId = user.Id;
+                newStore.CreatedAt = DateTime.UtcNow;
+
+                manager.stores.CreateStore(newStore);
+                await manager.Save();
+
+                var response = mapper.Map<StoresDto>(newStore);
+                return response;
+            }
+
+            throw new Exception("User already has a store");
         }
 
-        public async Task DeleteStore(Guid Id)
+        public async Task DeleteStore(string username, Guid Id)
         {
+            var user = await userManager.FindByNameAsync(username);
+
             var store = await manager.stores.GetStoreById(Id, false);
 
             if(store is null)
@@ -43,14 +63,25 @@ namespace Services
                 throw new StoreNotFoundException(Id);
             }
 
-            manager.stores.DeleteStore(store);
-            await manager.Save();
+            if(store.UserId == user.Id || await userManager.IsInRoleAsync(user, "Administrator"))
+            {
+                manager.stores.DeleteStore(store);
+                await manager.Save();
+            }
+            else
+            {
+                throw new Exception("Not authorized to delete this store");
+            }
 
         }
 
-        public async Task<StoresDto> GetStoreById(Guid id, bool trackChanges)
+        public async Task<StoresDto> GetStoreById(Guid id, string requester, bool trackChanges)
         {
+            var user = await userManager.FindByNameAsync(requester);
+
             var result = await manager.stores.GetStoreById(id, trackChanges);
+
+            bool requestedByUser = result.UserId.Equals(user.Id);
 
             if(result is null)
             {
@@ -58,21 +89,25 @@ namespace Services
             }
 
             var response = mapper.Map<StoresDto>(result);
+            response.requestedByOwner = requestedByUser;
             return response;
         }
 
-        public async Task<IEnumerable<StoresDto>> GetStores(bool trackChanges)
+        public async Task<(IEnumerable<StoresDto> stores, Metadata metadata)> GetStores(StoreParameters parameters, bool trackChanges)
         {
-            var results = await manager.stores.GetAllStores(trackChanges);
+            //Get Store Owners with stores
+
+            var results = await manager.stores.GetAllStores(parameters, trackChanges);
 
             var response = mapper.Map<IEnumerable<StoresDto>>(results);
 
-            return response;
-
+            return (stores: response, results.Metadata);
         }
 
-        public async Task UpdateStore(Guid Id, StoreUpdateDto updateModel, bool trackChanges)
+        public async Task UpdateStore(Guid Id, string username, StoreUpdateDto updateModel, bool trackChanges)
         {
+            var user = await userManager.FindByNameAsync(username);
+
             var store = await manager.stores.GetStoreById(Id, trackChanges);
 
             store.UpdatedAt = DateTime.Now;
@@ -82,9 +117,15 @@ namespace Services
                 throw new StoreNotFoundException(Id);
             }
 
-            mapper.Map(updateModel, store);
-
-            await manager.Save();
+            if (store.UserId == user.Id || await userManager.IsInRoleAsync(user, "Administrator"))
+            {
+                mapper.Map(updateModel, store);
+                await manager.Save();
+            }
+            else
+            {
+                throw new Exception("Not authorized to update this store");
+            }
         }
     }
 }
